+++ src/analytic/axisymmetric_tokamak_cartesian.jl
@@ -17,7 +17,6 @@ Parameters:
 """
 module AxisymmetricTokamakCartesian
 
-
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
     import ..AnalyticCartesianField: X, Y, Z