+++ src/analytic/quadratic_potentials.jl
@@ -18,7 +18,6 @@ and electrostatic potential
 """
 module QuadraticPotentials
 
-
 import ..ElectromagneticFields
 import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
 import ..AnalyticCartesianField: X, Y, Z