+++ src/analytic/singular.jl
@@ -17,7 +17,6 @@ Parameters: `B₀`
 """
 module Singular
 
-
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
     import ..ElectromagneticFields: A₁, A₂, A₃