+++ src/analytic/solovev_symmetric.jl
@@ -14,7 +14,6 @@ Parameters:
 """
 module SolovevSymmetric
 
-
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
     import ..AnalyticCartesianField: X, Y, Z